Move Nd4j tensor to separate module
